// lib/field-assets.generated.ts
export const FIELD_GROUND_ATLAS = {
  "width": 1280,
  "height": 448,
  "tiles": {
    "grass": {
      "x": 0,
      "y": 0,
      "width": 320,
      "height": 224
    },
    "dirt": {
      "x": 320,
      "y": 0,
      "width": 320,
      "height": 224
    },
    "paving": {
      "x": 640,
      "y": 0,
      "width": 320,
      "height": 224
    },
    "concrete": {
      "x": 960,
      "y": 0,
      "width": 320,
      "height": 224
    },
    "kampungGround": {
      "x": 0,
      "y": 224,
      "width": 320,
      "height": 224
    },
    "parkGrass": {
      "x": 320,
      "y": 224,
      "width": 320,
      "height": 224
    },
    "parkPaving": {
      "x": 640,
      "y": 224,
      "width": 320,
      "height": 224
    },
    "canalGrass": {
      "x": 960,
      "y": 224,
      "width": 320,
      "height": 224
    }
  }
} as const;
export type FieldAssetId = keyof typeof FIELD_OBJECT_ATLAS.assets;
export type FieldAnimatedId = keyof typeof FIELD_ANIMATED_ATLAS.animations;
export type GroundTileId = keyof typeof FIELD_GROUND_ATLAS.tiles;